backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-cloud-storage
firebase-admin
tensorflow-cpu
pillow
numpy
requests
